開放封閉原則的範例
Just as a recap, the open closed principle states that parts of a system or indeed subsystems have

to be open for extension.

So you should be able to extend the functionality of a filter, for example.

But they should be closed for modification.

So you shouldn't have to go back into beta filter, for example, and start adding things right here.

You shouldn't be able to do that.

Instead, if you want more functionality, you make new classes, you implement, I specification and

you feed those into something that has already been made and something that may have already been shipped.

You don't want to re ship the functionality of beta filter to your customers, but you can ship additional

modules which implement AI specification and which make use of better filter.

So that is what the open closed principle is all about.